+++ pybmix/core/hierarchy_mixing_changer.py
@@ -3,8 +3,4 @@ import os
 
 
 def change_hierarchy(new_name):
-    os.environ['HIER_C_NAME'] = new_name
-
-
-def change_non_conjugate_hierarchy(new_name):
-    os.environ['HIER_NC_NAME'] = new_name
+    os.environ['HIER_NAME'] = new_name
