+++ pybmix/core/pybmixcpp/bayesmix/py_hier/python_hierarchy.h
@@ -342,7 +342,7 @@ protected:
     std::vector<double> sum_stats;
 
     py::module_ numpy = py::module_::import("numpy");
-    py::module_ fun = py::module_::import("hierarchy_nc_implementation");
+    py::module_ fun = py::module_::import("hier_implementation");
     py::module_ numpy_random = py::module_::import("numpy.random");
     py::object py_engine = numpy_random.attr("MT19937")();
     py::object py_gen = numpy_random.attr("Generator")(py_engine);
